Announce chord duration with stem signs
This was a tough one, we had to implement decomposition of fractional
values into a list of augmented fractions first.
But we have a pretty robust system for expressing arbitrary values
now.
Additionally, whenever we see two chord symbols in a measure
we interleave with the duration expressed in stem sign symbols
with the utilities described above.